State field chain seems a constant
All uses of chain take `{var, chain}` as starting point in `_args`. This means that update in the `_next` function has no effect, it is reset to the original state.
This is weird in itself, but if it is just a pointer to the blockchain, then it makes perfect sense. But in that case, we explicitly treat it as constant.